+++ packages/terra-core-docs/src/terra-dev-site/doc/tag/example/TagList.jsx
@@ -6,8 +6,8 @@ import Tag from 'terra-tag';
 import TagList from 'terra-tag/src/TagList';
 import ThemeContext from 'terra-theme-context';
 import classNamesBind from 'classnames/bind';
-import styles from './FilterTagsCommon.module.scss';
 import Toolbar from 'terra-toolbar';
+import styles from './FilterTagsCommon.module.scss';
 
 const cx = classNamesBind.bind(styles);
 
@@ -17,39 +17,26 @@ const text = 'At vero eos et accusamus et iusto odio dignissimos ducimus qui bla
 const TagListExample = () => {
   const theme = React.useContext(ThemeContext);
 
-  const tagsData = [];
-  for (let tagCount = 1; tagCount < 11; tagCount += 1) {
-    const tagData = {
-      label: `Selection ${tagCount}`,
-      labelCategory: 'Category',
-      id: `terra-filter-tags-category-selection-${tagCount}`,
-    };
-    tagsData.push(tagData);
-  }
-
-  const [tags, setTags] = useState(tagsData);
-
   return (
     <>
       <div className={cx(['terra-filter-tags-doc-example-applied-filters', theme.className])}>
         <Toolbar className={cx('terra-filter-tags-doc-example-applied-filters-toolbar')}>
           <div className={cx('toolbar-flex-container')}>
-            <div className={cx('toolbar-flex-item-start')}>
-            </div>
+            <div className={cx('toolbar-flex-item-start')} />
             <div className={cx('toolbar-flex-item-fill')}>
-              <TagList isCollapsible aria-label={"blah blah"}>
-                <Tag icon={icon} text="Icon Tag" onClick={() => window.alert('Tag has been clicked!')} id="iconTag" />
+              <TagList isCollapsible aria-label='TagList Example'>
+                <Tag icon={icon} text="Icon Tag 1" onClick={() => window.alert('Tag has been clicked!')} id="iconTag1" />
                 <Tag icon={icon} text={text} id="longTextIcon" onClick={() => window.alert('Tag has been clicked!')} />
-                <Tag id="default" text="Default Tag" onClick={() => window.alert('Tag has been clicked!')} />
+                <Tag id="default1" text="Default Tag 1" onClick={() => window.alert('Tag has been clicked!')} />
                 <Tag id="on-click" text="Clickable Tag" onClick={() => window.alert('Tag has been clicked!')} />
-                <Tag icon={<IconTag />} text="Tag 5" onClick={() => window.alert('Tag has been clicked!')} id="tag1" />
-                <Tag icon={icon} text="Tag 6" onClick={() => window.alert('Tag has been clicked!')} id="tag2" />
-                <Tag text="Tag 7" onClick={() => window.alert('Tag has been clicked!')} id="tag3" />
-                <Tag icon={<IconTag />} text="Tag 8" onClick={() => window.alert('Tag has been clicked!')} id="tag4" />
-                <Tag icon={icon} text="Tag 9" onClick={() => window.alert('Tag has been clicked!')} id="tag5" />
-                <Tag text="Tag 10" onClick={() => window.alert('Tag has been clicked!')} id="tag6" />
-                <Tag icon={icon} text="Tag 11" onClick={() => window.alert('Tag has been clicked!')} id="tag7" />
-                <Tag icon={<IconTag />} text="Tag 12" onClick={() => window.alert('Tag has been clicked!')} id="tag8" />
+                <Tag icon={<IconTag />} text="Icon Tag 2" onClick={() => window.alert('Tag has been clicked!')} id="iconTag2" />
+                <Tag icon={icon} text="Icon Tag 3" onClick={() => window.alert('Tag has been clicked!')} id="iconTag3" />
+                <Tag text="Default Tag 2" onClick={() => window.alert('Tag has been clicked!')} id="default2" />
+                <Tag icon={<IconTag />} text="Icon Tag 4" onClick={() => window.alert('Tag has been clicked!')} id="iconTag4" />
+                <Tag icon={icon} text="Icon Tag 5" onClick={() => window.alert('Tag has been clicked!')} id="iconTag5" />
+                <Tag text="Default Tag 3" onClick={() => window.alert('Tag has been clicked!')} id="default3" />
+                <Tag icon={icon} text="Icon Tag 6" onClick={() => window.alert('Tag has been clicked!')} id="iconTag6" />
+                <Tag icon={<IconTag />} text="Icon Tag 7" onClick={() => window.alert('Tag has been clicked!')} id="iconTag7" />
               </TagList>
               <TagList isCollapsible>
                 <Tag href="http://google.com" text="HREF Tag" id="Tag1" />
@@ -58,8 +45,8 @@ const TagListExample = () => {
             </div>
           </div>
         </Toolbar>
-    </div>
-  </>
+      </div>
+    </>
   )
 };